// assets/games/tongits/script/proto/tongits.ts
export interface RoomInfo {
  roomId: number;
  roomName: string;
  /** 房间状态 */
  roomStatus: number;
  /** 最大座位数 */
  maxSeat: number;
}

/** 玩家在游戏中的完整信息 */
export interface TongitsPlayerInfo {
  /** 基础玩家信息 */
  playerInfo:
    | PlayerInfo
    | undefined;
  /** 手牌数量 (对其他玩家可见) */
  handCardCount: number;
  /** 是否是庄家 */
  isDealer: boolean;
  /** 已经打出的牌组 (明牌) */
  displayedMelds: Meld[];
  /** 自己的手牌 老 */
  handCards: number[];
  /** 是否可以发起挑战 */
  isFight: boolean;
  /** 时间戳 */
  countdown: number;
  /** 1:默认待选择 2:发起 3:接受 4:拒绝 5:烧死 */
  changeStatus: number;
  /** 1:不可操作, 2:select, 3:Action */
  status: number;
  /** 是否赢得比赛 */
  isWin: boolean;
  /** 点数 */
  cardPoint: number;
  /** 自动组牌 */
  isAuto: boolean;
  /** 自己的手牌 新 */
  groupCards: Cards[];
}

export interface TongitsBasePlayerInfo {
  /** 玩家ID */
  playerId: number;
  /** 1:默认待选择 2:发起 3:接受 4:拒绝 5:烧死 */
  changeStatus: number;
  /** 剩余时间 */
  countdown: number;
}

/** 牌组结构 */
export interface Meld {
  /** 组成牌组的牌 */
  cards: number[];
  /** 高亮的牌 (可见) */
  highlightCards: number;
  /** 这个牌组属于哪个玩家 */
  ownerId: number;
  /** 牌组的唯一ID (例如，玩家Melds数组的索引) */
  meldId: number;
  /** 牌组是否锁定 (动画表现) */
  locked: boolean;
}

/** 游戏状态的核心信息 */
export interface GameInfo {
  /** 当前操作玩家ID */
  actionPlayerId: number;
  /** 游戏阶段 1:初始 2:游戏中  3: 挑战中 4: 结算前  5:结算中 */
  status: number;
  /** 牌堆剩余牌数 */
  deckCardCount: number;
  /** 弃牌堆的牌 从下到上 */
  discardPile: number[];
  /** 底池信息 */
  pot:
    | PotInfo
    | undefined;
  /** 弃牌显示牌 */
  discardCard: number;
  /** 胜利类型  1 tongits 2 挑战 3 时间结束比大小: */
  winType: number;
  /** 视角玩家id */
  perspectiveId: number;
  /** 当前下注币值 */
  betAmount: number;
  /** 币值列表 */
  scores: number[];
}

/** 底池信息 */
export interface PotInfo {
  /** 底池金额 */
  base: number;
  /** 连续未被赢取的次数 */
  winCount: number;
  /** 当前所属玩家id */
  useId: number;
}
// ...
export interface Cards {
  /** 牌组id   0是散牌  1 ... 是牌组 */
  groupId: number;
  /** 牌组手牌 */
  handCards: number[];
  /** 牌型 0 无效牌型 1 有效牌型 2特殊牌型 */
  cardType: number;
  /** 牌组点数 */
  cardPoint: number;
}
// ...
export const Meld = {
  encode(message: Meld, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    writer.uint32(10).fork();
    for (const v of message.cards) {
      writer.int32(v);
    }
    writer.ldelim();
    if (message.highlightCards !== 0) {
      writer.uint32(16).int32(message.highlightCards);
    }
    if (message.ownerId !== 0) {
      writer.uint32(24).int64(message.ownerId);
    }
    if (message.meldId !== 0) {
      writer.uint32(32).int32(message.meldId);
    }
    if (message.locked !== false) {
      writer.uint32(40).bool(message.locked);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Meld {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMeld();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag === 8) {
            message.cards.push(reader.int32());

            continue;
          }

          if (tag === 10) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.cards.push(reader.int32());
            }

            continue;
          }

          break;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.highlightCards = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.ownerId = longToNumber(reader.int64() as Long);
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.meldId = reader.int32();
          continue;
        case 5:
          if (tag !== 40) {
            break;
          }

          message.locked = reader.bool();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Meld {
    return {
      cards: globalThis.Array.isArray(object?.cards) ? object.cards.map((e: any) => globalThis.Number(e)) : [],
      highlightCards: isSet(object.highlightCards) ? globalThis.Number(object.highlightCards) : 0,
      ownerId: isSet(object.ownerId) ? globalThis.Number(object.ownerId) : 0,
      meldId: isSet(object.meldId) ? globalThis.Number(object.meldId) : 0,
      locked: isSet(object.locked) ? globalThis.Boolean(object.locked) : false,
    };
  },

  toJSON(message: Meld): unknown {
    const obj: any = {};
    if (message.cards?.length) {
      obj.cards = message.cards.map((e) => Math.round(e));
    }
    if (message.highlightCards !== 0) {
      obj.highlightCards = Math.round(message.highlightCards);
    }
    if (message.ownerId !== 0) {
      obj.ownerId = Math.round(message.ownerId);
    }
    if (message.meldId !== 0) {
      obj.meldId = Math.round(message.meldId);
    }
    if (message.locked !== false) {
      obj.locked = message.locked;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Meld>, I>>(base?: I): Meld {
    return Meld.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Meld>, I>>(object: I): Meld {
    const message = createBaseMeld();
    message.cards = object.cards?.map((e) => e) || [];
    message.highlightCards = object.highlightCards ?? 0;
    message.ownerId = object.ownerId ?? 0;
    message.meldId = object.meldId ?? 0;
    message.locked = object.locked ?? false;
    return message;
  },
};
// ...
export const PotInfo = {
  encode(message: PotInfo, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.base !== 0) {
      writer.uint32(8).int32(message.base);
    }
    if (message.winCount !== 0) {
      writer.uint32(16).int32(message.winCount);
    }
    if (message.useId !== 0) {
      writer.uint32(24).int64(message.useId);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): PotInfo {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePotInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.base = reader.int32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.winCount = reader.int32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.useId = longToNumber(reader.int64() as Long);
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PotInfo {
    return {
      base: isSet(object.base) ? globalThis.Number(object.base) : 0,
      winCount: isSet(object.winCount) ? globalThis.Number(object.winCount) : 0,
      useId: isSet(object.useId) ? globalThis.Number(object.useId) : 0,
    };
  },

  toJSON(message: PotInfo): unknown {
    const obj: any = {};
    if (message.base !== 0) {
      obj.base = Math.round(message.base);
    }
    if (message.winCount !== 0) {
      obj.winCount = Math.round(message.winCount);
    }
    if (message.useId !== 0) {
      obj.useId = Math.round(message.useId);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PotInfo>, I>>(base?: I): PotInfo {
    return PotInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PotInfo>, I>>(object: I): PotInfo {
    const message = createBasePotInfo();
    message.base = object.base ?? 0;
    message.winCount = object.winCount ?? 0;
    message.useId = object.useId ?? 0;
    return message;
  },
};
// ...
export const Cards = {
  encode(message: Cards, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.groupId !== 0) {
      writer.uint32(8).int32(message.groupId);
    }
    writer.uint32(18).fork();
    for (const v of message.handCards) {
      writer.int32(v);
    }
    writer.ldelim();
    if (message.cardType !== 0) {
      writer.uint32(24).int32(message.cardType);
    }
    if (message.cardPoint !== 0) {
      writer.uint32(32).int32(message.cardPoint);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Cards {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCards();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.groupId = reader.int32();
          continue;
        case 2:
          if (tag === 16) {
            message.handCards.push(reader.int32());

            continue;
          }

          if (tag === 18) {
            const end2 = reader.uint32() + reader.pos;
            while (reader.pos < end2) {
              message.handCards.push(reader.int32());
            }

            continue;
          }

          break;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.cardType = reader.int32();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.cardPoint = reader.int32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Cards {
    return {
      groupId: isSet(object.groupId) ? globalThis.Number(object.groupId) : 0,
      handCards: globalThis.Array.isArray(object?.handCards)
        ? object.handCards.map((e: any) => globalThis.Number(e))
        : [],
      cardType: isSet(object.cardType) ? globalThis.Number(object.cardType) : 0,
      cardPoint: isSet(object.cardPoint) ? globalThis.Number(object.cardPoint) : 0,
    };
  },

  toJSON(message: Cards): unknown {
    const obj: any = {};
    if (message.groupId !== 0) {
      obj.groupId = Math.round(message.groupId);
    }
    if (message.handCards?.length) {
      obj.handCards = message.handCards.map((e) => Math.round(e));
    }
    if (message.cardType !== 0) {
      obj.cardType = Math.round(message.cardType);
    }
    if (message.cardPoint !== 0) {
      obj.cardPoint = Math.round(message.cardPoint);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Cards>, I>>(base?: I): Cards {
    return Cards.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Cards>, I>>(object: I): Cards {
    const message = createBaseCards();
    message.groupId = object.groupId ?? 0;
    message.handCards = object.handCards?.map((e) => e) || [];
    message.cardType = object.cardType ?? 0;
    message.cardPoint = object.cardPoint ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };